Moved reused county-buttons and search from js/equity to js/common.
