+++ src/js/charts-sandbox/index.js
@@ -1,8 +1,8 @@
 console.log("Loading chart sandbox");
 // import './charts/cagov-chart-vaccines-hpi-by-people/index.js'
 // import './charts/cagov-chart-vaccines-hpi-by-dose/index.js'
-import '../equity-dash/search/index.js';
-import '../equity-dash/county-buttons/buttons.js'
+import '../common/search/index.js';
+import '../common/county-buttons/buttons.js'
 
 import "../dashboard-v3/charts/cagov-chart-dashboard-confirmed-cases/index.js"
 
@@ -16,4 +16,4 @@ import "../dashboard-v3/charts/cagov-chart-dashboard-positivity-rate/index.js"
 import "../dashboard-v3/charts/cagov-chart-dashboard-patients/index.js"
 import "../dashboard-v3/charts/cagov-chart-dashboard-icu-beds/index.js"
 
-import '../equity-dash/search/filters.js'
+import '../common/search/filters.js'
