+++ src/js/common/search/index.js
