+++ src/js/common/search/readme.md
