+++ src/js/common/search/template.js
