+++ src/js/equity-dash/index.js
@@ -1,6 +1,6 @@
-import './search/index.js'
-import './county-buttons/buttons.js'
-import './search/filters.js'
+import '../common/search/index.js'
+import '../common/county-buttons/buttons.js'
+import '../common/search/filters.js'
 import './charts/relative-percentage-by-100k/index.js'
 import './charts/relative-percentage-by-pop/index.js'
 import './charts/healthy-places-index/index.js'
